Deployment changes
I changed the way I deploy my app as there was a lot of cache issues with the angular frontend.
This method will ask for a manual rebuild of the app each time the code change, but at least it will be done and there won't be weird cache things.